wip: splitting gobler, design finalization
